+++ src/components/base/input/InputSelect.vue
@@ -2,8 +2,8 @@
 	<Select.Root v-bind="root" v-model="model">
 		<Select.Trigger
 			v-bind="trigger"
-			class="inline-flex max-w-full items-center justify-between gap-1 rounded border
-				border-ui-menu-outline bg-ui-menu-inner px-1 py-0.5 text-ui-menu-text"
+			class="inline-flex max-w-full items-center justify-between gap-1 rounded-ui-select border
+				border-ui-select-outline bg-ui-select-bg px-1 py-0.5 text-ui-select-text"
 			:class="$attrs.class"
 		>
 			<Select.Value :placeholder class="h-[1.5em] truncate" />
@@ -13,8 +13,8 @@
 		<Select.Portal disabled>
 			<Select.Content
 				v-bind="content"
-				class="z-50 rounded border border-ui-menu-bg-outline bg-ui-menu-bg-inner p-0.5
-					text-ui-menu-text select-none"
+				class="z-50 rounded-ui-select-menu border border-ui-select-menu-outline bg-ui-select-menu-bg
+					p-0.5 text-ui-menu-text select-none"
 			>
 				<ScrollContainer>
 					<Select.Viewport class="flex flex-col gap-1">
@@ -23,8 +23,10 @@
 							v-bind="item"
 							:key="'option_' + val.label"
 							:value="val.value"
-							class="rounded p-1 text-xs data-highlighted:text-ui-menu-item-selected
-								data-highlighted:bg-gray-500 data-[state='checked']:bg--menu-item-inner-selected"
+							class="rounded-ui-select p-1 text-xs data-highlighted:brightness-125
+								data-highlighted:bg-ui-select-bg-selected
+								data-[state='checked']:bg-ui-select-bg-selected
+								data-[state='checked']:text-ui-select-text-selected cursor-pointer"
 						>
 							<MxTooltip :tooltip-disabled="!val.tooltip" :tooltip="val.tooltip">
 								<Select.ItemText>{{ val.label }}</Select.ItemText>